crime_extraction_service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
motor==3.3.2
azure-cosmos==4.5.1
httpx==0.26.0
ollama==0.1.6
python-dotenv==1.0.0
structlog==24.1.0
